poking and truncating
the minimal interface is now completed. all operations first check the
commands whether there is no garbage before executing the operation. the
number parsing code did not properly check whether the characters were
limited in the provided base (e.g. using hexadecimal `a' in decimal
mode).
the interpreter line buffer is expanded and redefined as a macro for
easier maintenance. this should be sufficient for most use cases.