split plain code, bfile_open redesign
opening the binary file has been redesigned to match proper behavior. if
a file size has been specified, it is implied that the file must be
created exclusively. therefore, if it can't be created, the operation
fails. if no file size has been specified, it is implied that the
specified file must exist.

also, partial opening of the file resulted in leaking a file descriptor,
this has been fixed by keeping track of the error and always checking
whether the error handler must be run. see also the documentation in
fs.h and implementation in bfile.c for more details.