add mapping for zero length files
Zero length files could not be opened properly because mmap and mremap
do not allow zero length mappings. This patch addresses this problem and
adds some bookkeeping to make sure a mapping is opened if the file is
truncated to a non-zero length.

The plain command parser has undergone some changes as well to be a
little more strict if a command has trailing garbage.